move constants out of /config
