+++ src/js/actions/actions-items.js
@@ -1,6 +1,6 @@
 import { itemsService } from "../services/service-items";
 import { dispatcher } from "../lib/game-dispatcher";
-import * as constants from "../config/constants-actions.json";
+import * as constants from "../constants/constants-actions.json";
 
 export const loadItems = levelName => {
   return new Promise((resolve, reject) => {
