+++ src/js/constants/combat.js
@@ -1,5 +1,5 @@
-{
+export default {
   "DMG_PROTECTED_MOD": 0.65,
   "DMG_VULNERABLE_MOD": 1.5,
   "COMBAT_ACTION_ATTACK": "COMBAT_ACTION_ATTACK"
-}
+};
