+++ src/js/constants/index.js
@@ -1,5 +1,5 @@
-{
+export default {
   "MSG_SPEED_MED": 500,
   "CHARACTER": "CHARACTER",
   "OPPONENT": "OPPONENT"
-}
+};
