+++ src/js/stores/store-inventory.js
@@ -2,7 +2,7 @@ import cloneDeep from "lodash.cloneDeep";
 
 import Store from "../lib/store";
 import { dispatcher } from "../lib/game-dispatcher";
-import * as constants from "../config/constants-actions";
+import constants from "../constants/actions";
 
 import ItemFactory from "../lib/item-factory";
 import Weapon from "../models/model-weapon";
