+++ src/js/stores/store-items.js
@@ -1,6 +1,6 @@
 import Store from "../lib/store";
 import { dispatcher } from "../lib/game-dispatcher";
-import * as constants from "../config/constants-actions";
+import constants from "../constants/actions";
 import ItemFactory from "../lib/item-factory";
 
 class ItemStore extends Store {
