+++ src/js/stores/store-level.js
@@ -3,7 +3,7 @@ import cloneDeep from "lodash.cloneDeep";
 
 import Store from "../lib/store";
 import { dispatcher } from "../lib/game-dispatcher";
-import * as constants from "../config/constants-actions";
+import constants from "../constants/actions";
 import { TileFactory } from "../lib/tile-factory";
 import { Tile } from "../models/model-tile";
